Ajout de tests unitaires.
