+++ tests/Formatter/ObjectWithToString.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace Kronos\Tests\Log\Formatter;
+
+class ObjectWithToString
+{
+
+    /**
+     * @var string
+     */
+    public $property;
+
+    public function __toString()
+    {
+        return $this->property;
+    }
+}
