+++ tests/Formatter/ObjectWithoutToString.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Kronos\Tests\Log\Formatter;
+
+class ObjectWithoutToString
+{
+    public $property;
+}
